chore(cli): #1105 PR 3 — move compile.rs helpers into compile/ submodules
Pure file moves, no behavior change. Final PR for #1105 (stacked on PR 2).

PR 1 added helper functions inside compile.rs. PR 2 added per-target
bundle writers. PR 3 lifts those helpers and bundle writers into focused
modules under crates/perry/src/commands/compile/:

  compile/bundle_apple.rs (450 lines)
    bundle_for_watchos, bundle_for_tvos, bundle_for_visionos,
    private read_visionos_app_metadata

  compile/i18n_emit.rs (151 lines)
    write_i18n_key_registry, emit_android_i18n_resources,
    write_lproj_localized_strings

  compile/post_link.rs (124 lines)
    strip_final_binary, emit_attestation_sidecar, print_binary_size,
    cleanup_intermediates, summarize_codegen_cache_stats

  compile/resources.rs (65 lines)
    find_project_root_for_resources, copy_dir_recursive,
    copy_bundle_resource_dirs

Each module follows the existing compile/* convention: pub(super)
visibility, parent types via super::, anyhow Result. No public-API
change — run, run_with_parse_cache, CompileResult, CompileArgs,
CompilationContext all still live in compile.rs.

The CompilePipeline struct from the original PR 2 spec is explicitly
punted: it needs data-flow analysis across the remaining ~6940-line
frame and should land as a separate, deliberate PR. PR 3 has handed the
next refactorer named functions to thread through the struct.

Metadata (version bump + CHANGELOG entry) intentionally omitted — per
CLAUDE.md the maintainer folds those in at merge time.

Validation: cargo build --release -p perry clean, 332 perry tests pass,
host smoke + --target ios-simulator both produce baseline output.
