+++ crates/perry/src/commands/compile.rs
@@ -20,14 +20,24 @@ mod library_search;
 mod link;
 mod object_cache;
 mod optimized_libs;
+mod bundle_apple;
+mod i18n_emit;
 mod parse_cache;
+mod post_link;
 mod resolve;
+mod resources;
 mod sandbox_buildrs;
 mod strip_dedup;
 mod targets;
 pub mod well_known;
 mod widget_build;
+use bundle_apple::{bundle_for_tvos, bundle_for_visionos, bundle_for_watchos};
 use collect_modules::collect_modules;
+use i18n_emit::{emit_android_i18n_resources, write_i18n_key_registry};
+use post_link::{
+    cleanup_intermediates, emit_attestation_sidecar, print_binary_size, strip_final_binary,
+    summarize_codegen_cache_stats,
+};
 pub use library_search::find_library;
 pub(crate) use library_search::host_target_triple;
 use library_search::{
@@ -54,7 +64,7 @@ use targets::{
     apple_sdk_version, compile_for_android_widget, compile_for_ios_widget, compile_for_wasm,
     compile_for_watchos_widget, compile_for_wearos_tile, compile_metallib_for_bundle,
     find_visionos_swift_runtime, find_watchos_swift_runtime, generate_embedded_js_object,
-    generate_js_bundle, lookup_bundle_id_from_toml,
+    generate_js_bundle,
 };
 
 /// Result of a successful compilation
@@ -1221,84 +1231,6 @@ fn write_audit_manifest_logging_failures(ctx: &CompilationContext, format: Outpu
     }
 }
 
-/// Strip debug symbols from the final binary (reduces size significantly).
-/// Skipped for: dylib output, every cross-compilation target whose host
-/// `strip` can't parse foreign object formats (iOS/visionOS/tvOS/watchOS/
-/// HarmonyOS/Android), and when `PERRY_DEBUG_SYMBOLS=1` is set so crash
-/// backtraces stay symbolicated.
-///
-/// When `ctx.needs_plugins` is true the build uses `strip -x` to retain
-/// exported symbols — `dlopen`'d plugins resolve `hone_host_api_*` from
-/// the main executable's symbol table.
-#[allow(clippy::too_many_arguments)]
-fn strip_final_binary(
-    ctx: &CompilationContext,
-    exe_path: &Path,
-    target: Option<&str>,
-    is_dylib: bool,
-    is_ios: bool,
-    is_visionos: bool,
-    is_tvos: bool,
-    is_watchos: bool,
-    is_harmonyos: bool,
-) {
-    if is_dylib
-        || is_ios
-        || is_visionos
-        || is_tvos
-        || is_watchos
-        || is_harmonyos
-        || target == Some("android")
-        || std::env::var("PERRY_DEBUG_SYMBOLS").is_ok()
-    {
-        return;
-    }
-    if ctx.needs_plugins {
-        let _ = std::process::Command::new("strip")
-            .arg("-x")
-            .arg(exe_path)
-            .status();
-    } else {
-        let _ = std::process::Command::new("strip").arg(exe_path).status();
-    }
-}
-
-/// #504: emit `<binary>.attest.json` AFTER strip/codesign so the captured
-/// SHA-256 matches what users will actually download. Best-effort —
-/// errors log and continue.
-fn emit_attestation_sidecar(ctx: &CompilationContext, exe_path: &Path, format: OutputFormat) {
-    if !ctx.emit_attest {
-        return;
-    }
-    match super::attest::build_attestation(exe_path, &ctx.project_root) {
-        Ok(manifest) => match super::attest::write_attestation(exe_path, &manifest) {
-            Ok(sidecar) => {
-                if let OutputFormat::Text = format {
-                    println!("Wrote attestation: {}", sidecar.display())
-                }
-            }
-            Err(e) => {
-                if let OutputFormat::Text = format {
-                    eprintln!("warning: failed to write attestation: {}", e)
-                }
-            }
-        },
-        Err(e) => {
-            if let OutputFormat::Text = format {
-                eprintln!("warning: failed to build attestation: {}", e)
-            }
-        }
-    }
-}
-
-fn print_binary_size(format: OutputFormat, exe_path: &Path) {
-    if let OutputFormat::Text = format {
-        if let Ok(meta) = fs::metadata(exe_path) {
-            let size_mb = meta.len() as f64 / 1_048_576.0;
-            println!("Binary size: {:.1}MB", size_mb);
-        }
-    }
-}
 
 /// Collect each `--bundle-extensions` entry into the same
 /// `CompilationContext` as the user entry. Returns `(canonical_path,
@@ -1431,654 +1363,9 @@ fn apply_geisterhand_args(args: &CompileArgs, ctx: &mut CompilationContext) {
     }
 }
 
-/// i18n: write the key registry to `.perry/i18n-keys.json` for tooling
-/// (translators, key-extraction scripts) that needs a flat list of
-/// emitted keys + their string-table indices. Best-effort.
-fn write_i18n_key_registry(
-    ctx: &CompilationContext,
-    i18n_table: Option<&perry_transform::i18n::I18nStringTable>,
-) {
-    let table = match i18n_table {
-        Some(t) if !t.keys.is_empty() => t,
-        _ => return,
-    };
-    let perry_dir = ctx.project_root.join(".perry");
-    let _ = fs::create_dir_all(&perry_dir);
-    let registry: Vec<serde_json::Value> = table
-        .keys
-        .iter()
-        .enumerate()
-        .map(|(i, key)| {
-            serde_json::json!({
-                "key": key,
-                "string_idx": i,
-            })
-        })
-        .collect();
-    let registry_json = serde_json::json!({ "keys": registry });
-    let _ = fs::write(
-        perry_dir.join("i18n-keys.json"),
-        serde_json::to_string_pretty(&registry_json).unwrap_or_default(),
-    );
-}
 
-/// Remove intermediate `.o` files unless `--keep-intermediates` was passed.
-fn cleanup_intermediates(keep_intermediates: bool, obj_paths: &[PathBuf]) {
-    if keep_intermediates {
-        return;
-    }
-    for obj_path in obj_paths {
-        let _ = fs::remove_file(obj_path);
-    }
-}
 
-/// Bundle the object cache's hit/miss/store counters for the
-/// `CompileResult` return value. `None` when the cache was disabled
-/// (`--no-cache`, `PERRY_NO_CACHE=1`, or bitcode-link mode).
-fn summarize_codegen_cache_stats(
-    object_cache: &ObjectCache,
-) -> Option<(usize, usize, usize, usize)> {
-    if !object_cache.is_enabled() {
-        return None;
-    }
-    Some((
-        object_cache.hits(),
-        object_cache.misses(),
-        object_cache.stores(),
-        object_cache.store_errors(),
-    ))
-}
 
-/// Walk up from `start` looking for a project anchor (`package.json`
-/// or, for the watchOS case, `perry.toml`). Bounded to 5 levels so a
-/// runaway walk can't traverse the filesystem. Returns the deepest
-/// directory that holds an anchor; if none found within the bound,
-/// returns the starting input unchanged.
-fn find_project_root_for_resources(start: &Path, watch_for_perry_toml: bool) -> PathBuf {
-    let mut project_root = start.to_path_buf();
-    for _ in 0..5 {
-        if project_root.join("package.json").exists() {
-            break;
-        }
-        if watch_for_perry_toml && project_root.join("perry.toml").exists() {
-            break;
-        }
-        if let Some(parent) = project_root.parent() {
-            project_root = parent.to_path_buf();
-        } else {
-            break;
-        }
-    }
-    project_root
-}
-
-/// Recursive copy used by the per-target bundle writers. Mirrors the
-/// inline `copy_dir_recursive_standalone` / `copy_dir_recursive`
-/// helpers that existed in each bundle branch.
-fn copy_dir_recursive(src: &Path, dst: &Path) -> std::io::Result<()> {
-    fs::create_dir_all(dst)?;
-    for entry in fs::read_dir(src)? {
-        let entry = entry?;
-        let ty = entry.file_type()?;
-        let dest_path = dst.join(entry.file_name());
-        if ty.is_dir() {
-            copy_dir_recursive(&entry.path(), &dest_path)?;
-        } else {
-            fs::copy(entry.path(), &dest_path)?;
-        }
-    }
-    Ok(())
-}
-
-/// Copy the `logo` / `assets` / `resources` / `images` directories
-/// from `project_root` into `dest_dir`. Used by the bundle writers
-/// so `[[NSBundle mainBundle] resourcePath]` / `resolve_asset_path`
-/// can find assets at runtime.
-fn copy_bundle_resource_dirs(project_root: &Path, dest_dir: &Path) {
-    for dir_name in &["logo", "assets", "resources", "images"] {
-        let resource_dir = project_root.join(dir_name);
-        if resource_dir.is_dir() {
-            let dest = dest_dir.join(dir_name);
-            let _ = copy_dir_recursive(&resource_dir, &dest);
-        }
-    }
-}
-
-/// Create a watchOS `.app` bundle: copy the linked binary into the
-/// bundle, write `Info.plist` (CFBundleExecutable / CFBundleIdentifier
-/// / WKApplication), copy project asset directories, then run the
-/// shared metallib compile for any `.metal` sources.
-///
-/// Returns `(app_dir, bundle_id)` so the caller can wire them into
-/// the final `CompileResult` (`result_app_dir`, `result_bundle_id`).
-fn bundle_for_watchos(
-    exe_path: &Path,
-    stem: &str,
-    target: Option<&str>,
-    input: &Path,
-    ctx: &CompilationContext,
-    format: OutputFormat,
-) -> Result<(PathBuf, String)> {
-    let app_dir = exe_path.with_extension("app");
-    let _ = fs::create_dir_all(&app_dir);
-    let bundle_exe = app_dir.join(exe_path.file_name().unwrap_or_default());
-    fs::copy(exe_path, &bundle_exe)?;
-    let _ = fs::remove_file(exe_path);
-
-    let exe_stem = exe_path
-        .file_stem()
-        .and_then(|s| s.to_str())
-        .unwrap_or(stem);
-    let bundle_id = lookup_bundle_id_from_toml(input, "watchos")
-        .or_else(|| lookup_bundle_id_from_toml(input, "app"))
-        .unwrap_or_else(|| crate::commands::sanitize::default_perry_bundle_id(exe_stem));
-
-    let info_plist = format!(
-        r#"<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>CFBundleExecutable</key>
-    <string>{exe_stem}</string>
-    <key>CFBundleIdentifier</key>
-    <string>{bundle_id}</string>
-    <key>CFBundleName</key>
-    <string>{exe_stem}</string>
-    <key>CFBundleVersion</key>
-    <string>1.0</string>
-    <key>CFBundleShortVersionString</key>
-    <string>1.0</string>
-    <key>MinimumOSVersion</key>
-    <string>10.0</string>
-    <key>UIDeviceFamily</key>
-    <array>
-        <integer>4</integer>
-    </array>
-    <key>WKApplication</key>
-    <true/>
-    <key>WKWatchOnly</key>
-    <true/>
-</dict>
-</plist>"#
-    );
-    fs::write(app_dir.join("Info.plist"), info_plist)?;
-
-    // Copy project resource directories into the bundle so
-    // bloom_load_texture / load_sound / read_file can resolve relative
-    // asset paths via [[NSBundle mainBundle] resourcePath].
-    if let Some(src_dir) = input
-        .canonicalize()
-        .ok()
-        .and_then(|p| p.parent().map(|d| d.to_path_buf()))
-    {
-        let project_root = find_project_root_for_resources(&src_dir, true);
-        copy_bundle_resource_dirs(&project_root, &app_dir);
-    }
-
-    compile_metallib_for_bundle(ctx, target, &app_dir, format)?;
-
-    match format {
-        OutputFormat::Text => {
-            println!("Wrote watchOS app bundle: {}", app_dir.display());
-            println!();
-            println!("To run on Apple Watch Simulator:");
-            println!("  xcrun simctl install booted {}", app_dir.display());
-            println!("  xcrun simctl launch booted {}", bundle_id);
-        }
-        OutputFormat::Json => {
-            let result = serde_json::json!({
-                "success": true,
-                "output": app_dir.to_string_lossy(),
-                "bundle_id": bundle_id,
-                "native_modules": ctx.native_modules.len(),
-                "js_modules": ctx.js_modules.len(),
-            });
-            println!("{}", serde_json::to_string(&result)?);
-        }
-    }
-
-    Ok((app_dir, bundle_id))
-}
-
-/// Walk up from the input file looking for `perry.toml` and harvest
-/// visionOS-relevant metadata: project version + build number, the
-/// `[visionos] deployment_target` (falling back to `minimum_version`,
-/// or `"1.0"`), the optional `encryption_exempt` flag, and any
-/// `[visionos.info_plist]` key/value pairs serialized into a plist
-/// fragment ready to splice into the dict.
-///
-/// Returns sensible defaults when no `perry.toml` is found.
-fn read_visionos_app_metadata(input: &Path) -> (String, String, String, Option<bool>, String) {
-    let walk = (|| -> Option<(String, String, String, Option<bool>, String)> {
-        let mut dir = input.canonicalize().ok()?;
-        for _ in 0..5 {
-            dir = dir.parent()?.to_path_buf();
-            let toml_path = dir.join("perry.toml");
-            if !toml_path.exists() {
-                continue;
-            }
-            let data = fs::read_to_string(&toml_path).ok()?;
-            let doc: toml::Table = data.parse().ok()?;
-            let project = doc.get("project").and_then(|v| v.as_table());
-            let visionos = doc.get("visionos").and_then(|v| v.as_table());
-            let version = project
-                .and_then(|p| p.get("version"))
-                .and_then(|v| v.as_str())
-                .unwrap_or("1.0.0")
-                .to_string();
-            let build_number = project
-                .and_then(|p| p.get("build_number"))
-                .and_then(|v| {
-                    v.as_integer()
-                        .map(|n| n.to_string())
-                        .or_else(|| v.as_str().map(|s| s.to_string()))
-                })
-                .unwrap_or_else(|| "1".to_string());
-            let deployment_target = visionos
-                .and_then(|v| {
-                    v.get("deployment_target")
-                        .or_else(|| v.get("minimum_version"))
-                })
-                .and_then(|v| v.as_str())
-                .unwrap_or("1.0")
-                .to_string();
-            let encryption_exempt = visionos
-                .and_then(|v| v.get("encryption_exempt"))
-                .and_then(|v| v.as_bool());
-            let mut entries = String::new();
-            if let Some(info_plist) = visionos
-                .and_then(|v| v.get("info_plist"))
-                .and_then(|v| v.as_table())
-            {
-                for (key, value) in info_plist {
-                    if let Some(s) = value.as_str() {
-                        entries.push_str(&format!(
-                            "    <key>{}</key>\n    <string>{}</string>\n",
-                            key, s
-                        ));
-                    } else if let Some(b) = value.as_bool() {
-                        entries.push_str(&format!(
-                            "    <key>{}</key>\n    <{}/>\n",
-                            key,
-                            if b { "true" } else { "false" }
-                        ));
-                    } else if let Some(i) = value.as_integer() {
-                        entries.push_str(&format!(
-                            "    <key>{}</key>\n    <integer>{}</integer>\n",
-                            key, i
-                        ));
-                    }
-                }
-            }
-            return Some((
-                version,
-                build_number,
-                deployment_target,
-                encryption_exempt,
-                entries,
-            ));
-        }
-        Some((
-            "1.0.0".to_string(),
-            "1".to_string(),
-            "1.0".to_string(),
-            None,
-            String::new(),
-        ))
-    })();
-    walk.unwrap_or_else(|| {
-        (
-            "1.0.0".to_string(),
-            "1".to_string(),
-            "1.0".to_string(),
-            None,
-            String::new(),
-        )
-    })
-}
-
-/// Write `<bundle>/<locale>.lproj/Localizable.strings` for every
-/// configured locale. Used by the iOS / macOS / visionOS bundle
-/// writers — Foundation `NSLocalizedString(key)` resolves from these
-/// files when the user's preferred language matches the directory.
-fn write_lproj_localized_strings(
-    app_dir: &Path,
-    i18n_table: Option<&perry_transform::i18n::I18nStringTable>,
-    i18n_config: Option<&perry_transform::i18n::I18nConfig>,
-) {
-    let (table, config) = match (i18n_table, i18n_config) {
-        (Some(t), Some(c)) if !t.keys.is_empty() => (t, c),
-        _ => return,
-    };
-    for (locale_idx, locale) in config.locales.iter().enumerate() {
-        let lproj_dir = app_dir.join(format!("{}.lproj", locale));
-        let _ = fs::create_dir_all(&lproj_dir);
-        let mut strings_content = String::new();
-        for (key_idx, key) in table.keys.iter().enumerate() {
-            let flat_idx = locale_idx * table.keys.len() + key_idx;
-            let value = table
-                .translations
-                .get(flat_idx)
-                .cloned()
-                .unwrap_or_else(|| key.clone());
-            let escaped_key = key.replace('\\', "\\\\").replace('"', "\\\"");
-            let escaped_val = value.replace('\\', "\\\\").replace('"', "\\\"");
-            strings_content
-                .push_str(&format!("\"{}\" = \"{}\";\n", escaped_key, escaped_val));
-        }
-        let _ = fs::write(lproj_dir.join("Localizable.strings"), &strings_content);
-    }
-}
-
-/// Create a visionOS `.app` bundle. Reads `perry.toml` for version +
-/// build number + deployment target + encryption-exemption + custom
-/// Info.plist entries, copies the linked binary, writes Info.plist
-/// (with XRSimulator / XROS platform tag), copies project asset
-/// directories, and writes per-locale `.lproj/Localizable.strings`
-/// from the i18n table. Returns `(app_dir, bundle_id)`.
-fn bundle_for_visionos(
-    exe_path: &Path,
-    stem: &str,
-    target: Option<&str>,
-    input: &Path,
-    ctx: &CompilationContext,
-    i18n_table: Option<&perry_transform::i18n::I18nStringTable>,
-    i18n_config: Option<&perry_transform::i18n::I18nConfig>,
-    format: OutputFormat,
-) -> Result<(PathBuf, String)> {
-    let app_dir = exe_path.with_extension("app");
-    let _ = fs::create_dir_all(&app_dir);
-    let bundle_exe = app_dir.join(exe_path.file_name().unwrap_or_default());
-    fs::copy(exe_path, &bundle_exe)?;
-    let _ = fs::remove_file(exe_path);
-
-    let exe_stem = exe_path
-        .file_stem()
-        .and_then(|s| s.to_str())
-        .unwrap_or(stem);
-    let bundle_id = lookup_bundle_id_from_toml(input, "visionos")
-        .or_else(|| lookup_bundle_id_from_toml(input, "app"))
-        .or_else(|| lookup_bundle_id_from_toml(input, "ios"))
-        .unwrap_or_else(|| crate::commands::sanitize::default_perry_bundle_id(exe_stem));
-
-    let (app_version, app_build_number, deployment_target, encryption_exempt, custom_plist_entries) =
-        read_visionos_app_metadata(input);
-
-    let platform_name = if target == Some("visionos-simulator") {
-        "XRSimulator"
-    } else {
-        "XROS"
-    };
-
-    let mut info_plist = format!(
-        r#"<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>CFBundleExecutable</key>
-    <string>{exe_stem}</string>
-    <key>CFBundleIdentifier</key>
-    <string>{bundle_id}</string>
-    <key>CFBundleName</key>
-    <string>{exe_stem}</string>
-    <key>CFBundleVersion</key>
-    <string>{app_build_number}</string>
-    <key>CFBundleShortVersionString</key>
-    <string>{app_version}</string>
-    <key>CFBundlePackageType</key>
-    <string>APPL</string>
-    <key>CFBundleInfoDictionaryVersion</key>
-    <string>6.0</string>
-    <key>MinimumOSVersion</key>
-    <string>{deployment_target}</string>
-    <key>CFBundleSupportedPlatforms</key>
-    <array>
-        <string>{platform_name}</string>
-    </array>
-    <key>UIRequiredDeviceCapabilities</key>
-    <array>
-        <string>arm64</string>
-    </array>
-    <key>UIDeviceFamily</key>
-    <array>
-        <integer>7</integer>
-    </array>
-    <key>UILaunchScreen</key>
-    <dict/>
-    <key>UIApplicationSceneManifest</key>
-    <dict>
-        <key>UIApplicationSupportsMultipleScenes</key>
-        <true/>
-        <key>UIApplicationPreferredDefaultSceneSessionRole</key>
-        <string>UIWindowSceneSessionRoleApplication</string>
-        <key>UISceneConfigurations</key>
-        <dict/>
-    </dict>
-</dict>
-</plist>"#
-    );
-
-    let usage_descriptions = concat!(
-        "    <key>NSCameraUsageDescription</key>\n",
-        "    <string>This app uses the camera to identify colors.</string>\n",
-        "    <key>NSMicrophoneUsageDescription</key>\n",
-        "    <string>This app uses the microphone to measure sound levels.</string>\n",
-    );
-    info_plist = info_plist.replace(
-        "</dict>\n</plist>",
-        &format!("{}</dict>\n</plist>", usage_descriptions),
-    );
-
-    if let Some(exempt) = encryption_exempt {
-        let encryption_entry = format!(
-            "    <key>ITSAppUsesNonExemptEncryption</key>\n    <{}/>\n",
-            if exempt { "false" } else { "true" }
-        );
-        info_plist = info_plist.replace(
-            "</dict>\n</plist>",
-            &format!("{}</dict>\n</plist>", encryption_entry),
-        );
-    }
-
-    if !custom_plist_entries.is_empty() {
-        info_plist = info_plist.replace(
-            "</dict>\n</plist>",
-            &format!("{}</dict>\n</plist>", custom_plist_entries),
-        );
-    }
-
-    fs::write(app_dir.join("Info.plist"), info_plist)?;
-
-    if let Some(src_dir) = input
-        .canonicalize()
-        .ok()
-        .and_then(|p| p.parent().map(|d| d.to_path_buf()))
-    {
-        let project_root = find_project_root_for_resources(&src_dir, true);
-        copy_bundle_resource_dirs(&project_root, &app_dir);
-    }
-
-    write_lproj_localized_strings(&app_dir, i18n_table, i18n_config);
-
-    match format {
-        OutputFormat::Text => {
-            println!("Wrote visionOS app bundle: {}", app_dir.display());
-            println!();
-            println!("To run on Apple Vision Pro Simulator:");
-            println!("  xcrun simctl install booted {}", app_dir.display());
-            println!("  xcrun simctl launch booted {}", bundle_id);
-        }
-        OutputFormat::Json => {
-            let result = serde_json::json!({
-                "success": true,
-                "output": app_dir.to_string_lossy(),
-                "bundle_id": bundle_id,
-                "native_modules": ctx.native_modules.len(),
-                "js_modules": ctx.js_modules.len(),
-            });
-            println!("{}", serde_json::to_string(&result)?);
-        }
-    }
-
-    Ok((app_dir, bundle_id))
-}
-
-/// Create a tvOS `.app` bundle. Same shape as `bundle_for_watchos`
-/// but with the Apple TV `UIDeviceFamily` (3), `MinimumOSVersion`
-/// 17.0, and a UI-thread principal class of `BloomApplication` (the
-/// runtime's UIApplication subclass on tvOS). No project-resource
-/// copy in this path — tvOS apps consume assets via the metallib /
-/// embedded JS bundle.
-fn bundle_for_tvos(
-    exe_path: &Path,
-    stem: &str,
-    target: Option<&str>,
-    input: &Path,
-    ctx: &CompilationContext,
-    format: OutputFormat,
-) -> Result<(PathBuf, String)> {
-    let app_dir = exe_path.with_extension("app");
-    let _ = fs::create_dir_all(&app_dir);
-    let bundle_exe = app_dir.join(exe_path.file_name().unwrap_or_default());
-    fs::copy(exe_path, &bundle_exe)?;
-    let _ = fs::remove_file(exe_path);
-
-    let exe_stem = exe_path
-        .file_stem()
-        .and_then(|s| s.to_str())
-        .unwrap_or(stem);
-    let bundle_id = lookup_bundle_id_from_toml(input, "tvos")
-        .or_else(|| lookup_bundle_id_from_toml(input, "app"))
-        .unwrap_or_else(|| crate::commands::sanitize::default_perry_bundle_id(exe_stem));
-
-    let info_plist = format!(
-        r#"<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>CFBundleExecutable</key>
-    <string>{exe_stem}</string>
-    <key>CFBundleIdentifier</key>
-    <string>{bundle_id}</string>
-    <key>CFBundleName</key>
-    <string>{exe_stem}</string>
-    <key>CFBundleVersion</key>
-    <string>1.0</string>
-    <key>CFBundleShortVersionString</key>
-    <string>1.0</string>
-    <key>MinimumOSVersion</key>
-    <string>17.0</string>
-    <key>UIDeviceFamily</key>
-    <array>
-        <integer>3</integer>
-    </array>
-    <key>UILaunchScreen</key>
-    <dict/>
-    <key>UIRequiresFullScreen</key>
-    <true/>
-    <key>NSPrincipalClass</key>
-    <string>BloomApplication</string>
-</dict>
-</plist>"#
-    );
-    fs::write(app_dir.join("Info.plist"), info_plist)?;
-
-    compile_metallib_for_bundle(ctx, target, &app_dir, format)?;
-
-    match format {
-        OutputFormat::Text => {
-            println!("Wrote tvOS app bundle: {}", app_dir.display());
-            println!();
-            println!("To run on Apple TV Simulator:");
-            println!("  xcrun simctl install booted {}", app_dir.display());
-            println!("  xcrun simctl launch booted {}", bundle_id);
-        }
-        OutputFormat::Json => {
-            let result = serde_json::json!({
-                "success": true,
-                "output": app_dir.to_string_lossy(),
-                "bundle_id": bundle_id,
-                "native_modules": ctx.native_modules.len(),
-                "js_modules": ctx.js_modules.len(),
-            });
-            println!("{}", serde_json::to_string(&result)?);
-        }
-    }
-
-    Ok((app_dir, bundle_id))
-}
-
-/// i18n: emit `res/values-<locale>/strings.xml` next to the Android
-/// `.so` for every configured locale. Best-effort; failures don't
-/// abort the build. Keys are sanitized to `[A-Za-z0-9_]+` (Android
-/// resource-name grammar) and values get the usual XML entity escape.
-fn emit_android_i18n_resources(
-    is_android: bool,
-    i18n_table: Option<&perry_transform::i18n::I18nStringTable>,
-    i18n_config: Option<&perry_transform::i18n::I18nConfig>,
-    exe_path: &Path,
-    format: OutputFormat,
-) {
-    if !is_android {
-        return;
-    }
-    let (table, config) = match (i18n_table, i18n_config) {
-        (Some(t), Some(c)) => (t, c),
-        _ => return,
-    };
-    if table.keys.is_empty() {
-        return;
-    }
-    let output_dir = exe_path.parent().unwrap_or(Path::new("."));
-    let res_dir = output_dir.join("res");
-    for (locale_idx, locale) in config.locales.iter().enumerate() {
-        let values_dir = if locale_idx == 0 {
-            res_dir.join("values") // default locale
-        } else {
-            res_dir.join(format!("values-{}", locale))
-        };
-        let _ = fs::create_dir_all(&values_dir);
-        let mut xml =
-            String::from("<?xml version=\"1.0\" encoding=\"utf-8\"?>\n<resources>\n");
-        for (key_idx, key) in table.keys.iter().enumerate() {
-            let flat_idx = locale_idx * table.keys.len() + key_idx;
-            let value = table
-                .translations
-                .get(flat_idx)
-                .cloned()
-                .unwrap_or_else(|| key.clone());
-            let res_name: String = key
-                .chars()
-                .map(|c| {
-                    if c.is_alphanumeric() || c == '_' {
-                        c
-                    } else {
-                        '_'
-                    }
-                })
-                .collect();
-            let escaped = value
-                .replace('&', "&amp;")
-                .replace('<', "&lt;")
-                .replace('>', "&gt;")
-                .replace('"', "&quot;")
-                .replace('\'', "\\'");
-            xml.push_str(&format!(
-                "    <string name=\"{}\">{}</string>\n",
-                res_name, escaped
-            ));
-        }
-        xml.push_str("</resources>\n");
-        let _ = fs::write(values_dir.join("strings.xml"), &xml);
-    }
-    match format {
-        OutputFormat::Text => println!(
-            "  Generated res/values-*/strings.xml for {} locale(s)",
-            config.locales.len()
-        ),
-        OutputFormat::Json => {}
-    }
-}
 
 pub fn run(
     args: CompileArgs,
